refactor: rename buttons.js to config.js
