+++ index.html
@@ -8,7 +8,6 @@
   <title>Virtual Keyboard</title>
 </head>
 <body>
-  <script src="buttons.js" type="module"></script>
   <script src="main.js" type="module"></script>
 </body>
 </html>
